movie/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dependency-injector==4.44.0
Deprecated==1.2.15
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
injector==0.22.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
mirakuru==2.5.3
motor==3.6.0
MyApplication==0.1.0
mypy==1.13.0
mypy-extensions==1.0.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
pluggy==1.5.0
port-for==0.7.4
protobuf==5.29.1
psutil==6.1.0
psycopg==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.9.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-postgresql==6.1.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
